requirements.txt
opencv-python
numpy
ultralytics
Pillow
PyMuPDF